First real world proposed tests to the ToAE
Experimental prediction of the \gamma\kappa linear term for quantum measurements in gravitational setting, where a gravity term affects the measurement deviation. This will provide real clues to a real deviation with potential definitive values, or at least find upper limit bounds to search for it in other experiment. It is the first experiment out of the ToAE that can confirm its explanatory powers in physics.